Forest shouldn't have solve or solve_multiple, only iter_answers
