+++ chalk-solve/src/lib.rs
@@ -109,3 +109,4 @@ pub use solve::Guidance;
 pub use solve::Solution;
 pub use solve::Solver;
 pub use solve::SolverChoice;
+pub use solve::SubstitutionResult;
